Revamped controls with new player prefab: ArcherPlayer. Other changes in commit description.
Added player controls using PlayerInput and invocation of Unity events. This is contained in a new script named 'ArcherPlayer'. A new player prefab (made with a capsule) now has an aiming indicator that is a cube that floats in front of the player. Players will now respawn in the center of the level 1 when they fall off the stage - this feature only works on this level right now. Wind Hazards now have the option to affect player objects - default option is off so no changes to the current level design have been made. Dynamic camera was disabled to allow playtesting. Camera is now set back to orthographic view.